third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/function-details.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>  function firstLineFunction()

{
}

   function notFirstLineFunction()

{
}

var obj = {
    m: function() {}
}

function functionWithDisplayName() {}
functionWithDisplayName.displayName = "user-friendly name";

function functionWithDisplayNameGetter() {}
functionWithDisplayNameGetter.__defineGetter__("displayName", function() { return "FAIL_should_not_execute"; });

var smallClosure = (function(p) { return function() { return p; }; })("Capybara");

var bigClosure = (function(p) {
    var o = {
       e: 7,
       f: 5,
       get u() { return 3; },
       set v(value) { }
    };
    with (o) {
        try {
            throw Error("Test");
        } catch (ex) {
            return function() {
                return String(p) + String(ex) + u + e;
            };
        }
    }
})({});

function* gen() { yield [1,2,3] }

function test()
{
    function dumpFunctionDetails(properties)
    {
        var location = properties.get("[[FunctionLocation]]").value.value;
        InspectorTest.addResult("Function details: ");
        InspectorTest.addResult("lineNumber: " + location.lineNumber);
        InspectorTest.addResult("columnNumber: " + location.columnNumber);
        InspectorTest.addResult("scriptId is valid: " + !!location.scriptId);
        InspectorTest.addResult("functionName: " + properties.get("name").value.value);
        if (properties.has("[[IsGenerator]]"))
            InspectorTest.addResult("isGenerator: " + properties.get("[[IsGenerator]]").value.value);
    }

    function dumpFunctionNoScopes()
    {
        InspectorTest.addResult("scopeChain: n/a");
    }

    function dumpFunctionScope(pos, type, propertyDescriptors)
    {
        var variables;
        if (type == "Global") {
            variables = "<global object properties omitted>";
        } else {
            var varArray = [];
            for (var i = 0; i < propertyDescriptors.length; i++) {
                var d = propertyDescriptors[i];
                var valueStr;
                if (d.value) {
                    if (d.value.value)
                        valueStr = JSON.stringify(d.value.value);
                    else
                        valueStr = "<no string representation>";
                } else {
                    valueStr = "<no value>";
                }
                varArray.push(d.name + ": " + valueStr);
            }
            varArray.sort();
            variables = varArray.join();
        }
        InspectorTest.addResult("scopeChain #" + pos + ": " + type + "; " + variables);
    }

    function loadAndDumpScopeObjects(scopeChain, end)
    {
        var scopes = [];
        function loadScopeObject(pos, next)
        {
            if (pos >= scopes.length) {
                next();
                return;
            }
            InspectorTest.RuntimeAgent.getProperties(scopes[pos].objectId, true, didGetProperties);

            function didGetProperties(error, propertyDescriptors)
            {
                dumpFunctionScope(pos, scopes[pos].description, propertyDescriptors);
                loadScopeObject(pos + 1, next);
            }
        }

        if (scopeChain) {
            InspectorTest.RuntimeAgent.getProperties(scopeChain.value.objectId, true, didGetScopes);
        } else {
            dumpFunctionNoScopes();
            end();
        }

        function didGetScopes(error, properties)
        {
            for (var prop of properties) {
                if (String(prop.name >>> 0) === prop.name)
                    scopes.push(prop.value);
            }
            loadScopeObject(0, end);
        }
    }

    function performStandardTestCase(pageExpression, next)
    {
        InspectorTest.evaluateInPage(pageExpression, didEvaluate);

        function didEvaluate(remote)
        {
            InspectorTest.addResult(pageExpression + " type = " + remote.type);
            InspectorTest.RuntimeAgent.getProperties(remote.objectId, /* isOwnProperty */ false, didGetDetails);
        }
        function didGetDetails(error, properties, internalProperties)
        {
            var propertiesMap = new Map();
            for (var prop of internalProperties)
                propertiesMap.set(prop.name, prop);
            for (var prop of properties) {
                if (prop.name === "name" && prop.value && prop.value.type === "string")
                    propertiesMap.set("name", prop);
                if (prop.name === "displayName" && prop.value && prop.value.type === "string") {
                    propertiesMap.set("name", prop);
                    break;
                }
            }
            dumpFunctionDetails(propertiesMap);
            loadAndDumpScopeObjects(propertiesMap.get("[[Scopes]]"), next);
        }
    }

    InspectorTest.runDebuggerTestSuite([
        function testGetFirstLineFunctionDetails(next)
        {
            performStandardTestCase("firstLineFunction", next);
        },
        function testGetNonFirstLineFunctionDetails(next)
        {
            performStandardTestCase("notFirstLineFunction", next);
        },
        function testGetDetailsOfFunctionWithInferredName(next)
        {
            performStandardTestCase("obj.m", next);
        },
        function testGetDetailsOfFunctionWithDisplayName(next)
        {
            performStandardTestCase("functionWithDisplayName", next);
        },
        function testGetDetailsOfFunctionWithDisplayNameGetter(next)
        {
            performStandardTestCase("functionWithDisplayNameGetter", next);
        },
        function testSmallClosure(next)
        {
            performStandardTestCase("smallClosure", next);
        },
        function testBigClosure(next)
        {
            performStandardTestCase("bigClosure", next);
        },
        function testGenFunction(next)
        {
            performStandardTestCase("gen", next);
        }
    ]);
};

</script>
</head>

<body onload="runTest()">
<p>Tests that Debugger.getFunctionDetails command returns correct location.
<a href="https://bugs.webkit.org/show_bug.cgi?id=71808">Bug 71808</a>
</p>
</body>
</html>
