third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/function-details-expected.txt

Tests that Debugger.getFunctionDetails command returns correct location. Bug 71808


Running: testGetFirstLineFunctionDetails
firstLineFunction type = function
Function details: 
lineNumber: 5
columnNumber: 36
scriptId is valid: true
functionName: firstLineFunction
scopeChain #0: Global; <global object properties omitted>

Running: testGetNonFirstLineFunctionDetails
notFirstLineFunction type = function
Function details: 
lineNumber: 10
columnNumber: 32
scriptId is valid: true
functionName: notFirstLineFunction
scopeChain #0: Global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithInferredName
obj.m type = function
Function details: 
lineNumber: 16
columnNumber: 15
scriptId is valid: true
functionName: m
scopeChain #0: Global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithDisplayName
functionWithDisplayName type = function
Function details: 
lineNumber: 19
columnNumber: 32
scriptId is valid: true
functionName: user-friendly name
scopeChain #0: Global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithDisplayNameGetter
functionWithDisplayNameGetter type = function
Function details: 
lineNumber: 22
columnNumber: 38
scriptId is valid: true
functionName: functionWithDisplayNameGetter
scopeChain #0: Global; <global object properties omitted>

Running: testSmallClosure
smallClosure type = function
Function details: 
lineNumber: 25
columnNumber: 49
scriptId is valid: true
functionName: 
scopeChain #0: Closure; p: "Capybara"
scopeChain #1: Global; <global object properties omitted>

Running: testBigClosure
bigClosure type = function
Function details: 
lineNumber: 38
columnNumber: 27
scriptId is valid: true
functionName: 
scopeChain #0: Catch; ex: <no string representation>
scopeChain #1: With Block; __proto__: <no string representation>,e: 7,f: 5,u: <no value>,v: <no value>
scopeChain #2: Closure; p: <no string representation>
scopeChain #3: Global; <global object properties omitted>

Running: testGenFunction
gen type = function
Function details: 
lineNumber: 45
columnNumber: 13
scriptId is valid: true
functionName: gen
isGenerator: true
scopeChain #0: Global; <global object properties omitted>
